+++ examples/resources/growthbook_saved_group/import.sh
@@ -1,0 +1,1 @@
+terraform import growthbook_saved_group.beta_users sg_abc123